requirements.txt
flask>=2.3
groq>=0.13
python-dotenv>=1.1
Pillow>=10.0
pytest>=7.0
playwright>=1.40
